moved page registry to a dedicated module
